ERI-526 Atualizar fluxo do login
