+++ src/app/Certvet.tsx
@@ -10,7 +10,6 @@ import "@fortawesome/fontawesome-free/css/solid.min.css";
 import FormProntuario from "./views/form-prontuario/FormProntuario";
 import FormHistoricoClinico from "./views/form-historico-clinico/FormHistoricoClinico";
 import Agenda from "./views/agenda/Agenda";
-import Login from "./views/login/Login";
 import Painel from "./views/painel/Painel";
 import RedefinePassword from "./views/redefine-password/RedefinePassword";
 import Home from "./views/home/Home";
@@ -23,6 +22,7 @@ import Animals from "./views/animals/Animals";
 import SignIn from "./views/sign-in/SignIn";
 import NotFound from "./views/not-found/NotFound";
 import Medication from "./views/medication/Medication";
+import Auth from "./views/auth/Auth";
 
 export default function Certvet() {
     const [authoritesLoaded, setAuthoritesLoaded] = useState<boolean>(false);
@@ -49,7 +49,8 @@ export default function Certvet() {
         <BrowserRouter>
             <Routes>
                 <Route path="/" element={<Home/>}/>
-                <Route path="/login" element={<Login/>}/>
+                <Route path="/login" element={<Auth.Login/>}/>
+                <Route path="/login/redefinir" element={<Auth.ForgotPassword/>}/>
                 <Route path="/redefinir-senha" element={<RedefinePassword/>}/>
                 <Route path="/cadastro" element={<SignIn.Clinica/>}/>
                 <Route path="/cadastro/dono" element={<SignIn.Owner/>}/>
