+++ src/app/views/auth/Auth.ts
@@ -1,0 +1,9 @@
+import LoginPage from "./pages/login/Login";
+import ForgotPasswordPage from "./pages/forgot-password/ForgotPassword";
+
+namespace Auth {
+    export const Login = LoginPage;
+    export const ForgotPassword = ForgotPasswordPage;
+}
+
+export default Auth;
