+++ src/app/views/auth/pages/forgot-password/forgot-password.scss
@@ -1,0 +1,12 @@
+#forgot-password {
+  .rounded {
+    border-radius: 50px !important;
+  }
+
+  input {
+    border-top: none;
+    border-left: none;
+    border-right: none;
+    border-radius: 0;
+  }
+}
