--- src/app/views/login/login.scss
@@ -1,38 +1,0 @@
-#login {
-  .rounded {
-    border-radius: 50px !important;
-  }
-
-  input {
-    border-top: none;
-    border-left: none;
-    border-right: none;
-    border-radius: 0;
-  }
-
-  .clinicas-list {
-    max-height: 216px;
-    overflow-y: auto;
-
-    /* Firefox */
-    & {
-      scrollbar-width: auto;
-      scrollbar-color: #9c9c9c #ffffff;
-    }
-
-    /* Chrome, Edge, and Safari */
-    &::-webkit-scrollbar {
-      width: 10px;
-    }
-
-    &::-webkit-scrollbar-track {
-      background: #ffffff;
-    }
-
-    &::-webkit-scrollbar-thumb {
-      background-color: #9c9c9c;
-      border-radius: 10px;
-      border: 3px solid #ffffff;
-    }
-  }
-}
